refactor: split main file into smaller ones
